names -> releases::contributors
