+++ src/releases.rs
@@ -1,12 +1,24 @@
 use models::{Release, NewRelease};
+use models::Project;
+
+use caseless;
 
 use diesel;
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
 
+use serde_json::value::Value;
+
+use std::cmp::Ordering;
 use std::process::Command;
 
 use slog::Logger;
+
+use unicode_normalization::UnicodeNormalization;
+
+// needed for case-insensitivity
+use diesel::types::VarChar;
+sql_function!(lower, lower_t, (x: VarChar) -> VarChar);
 
 pub fn assign_commits(log: &Logger, release_name: &str, previous_release: &str, release_project_id: i32, path: &str) {
     // Could take the connection as a parameter, as problably
@@ -99,3 +111,90 @@ pub fn create(conn: &PgConnection, version: &str, project_id: i32) -> Release {
         .expect("Error saving new release")
 }
 
+pub fn contributors(project: &str, release_name: &str) -> Option<Vec<Value>> {
+    use schema::releases::dsl::*;
+    use schema::commits::dsl::*;
+    use models::Release;
+    use models::Commit;
+
+    let connection = ::establish_connection();
+
+    let project = {
+        use schema::projects::dsl::*;
+
+        match projects.filter(lower(name).eq(lower(project)))
+            .first::<Project>(&connection) {
+                Ok(p) => p,
+                Err(_) => {
+                    return None;
+                }
+        }
+    };
+
+    let release: Release = match releases
+        .filter(version.eq(release_name))
+        .filter(project_id.eq(project.id))
+        .first(&connection) {
+            Ok(release) => release,
+                Err(_) => {
+                    return None;
+                },
+        };
+
+    // it'd be better to do this in the db
+    // but Postgres doesn't do Unicode collation correctly on OSX
+    // http://postgresql.nabble.com/Collate-order-on-Mac-OS-X-text-with-diacritics-in-UTF-8-td1912473.html
+    let mut names: Vec<String> = Commit::belonging_to(&release)
+        .select(author_name).distinct().load(&connection).unwrap();
+
+    inaccurate_sort(&mut names);
+
+    Some(names.into_iter().map(Value::String).collect())
+}
+
+// TODO: switch this out for an implementation of the Unicode Collation Algorithm
+pub fn inaccurate_sort(strings: &mut Vec<String>) {
+    strings.sort_by(|a, b| str_cmp(&a, &b));
+}
+
+fn str_cmp(a_raw: &str, b_raw: &str) -> Ordering {
+    let a: Vec<char> = a_raw.nfkd().filter(|&c| (c as u32) < 0x300 || (c as u32) > 0x36f).collect();
+    let b: Vec<char> = b_raw.nfkd().filter(|&c| (c as u32) < 0x300 || (c as u32) > 0x36f).collect();
+
+    for (&a_char, &b_char) in a.iter().zip(b.iter()) {
+        match char_cmp(a_char, b_char) {
+            Ordering::Less => return Ordering::Less,
+            Ordering::Greater => return Ordering::Greater,
+            Ordering::Equal => {}
+        }
+    }
+
+    if a.len() < b.len() {
+        Ordering::Less
+    } else if a.len() > b.len() {
+        Ordering::Greater
+    } else {
+        Ordering::Equal
+    }
+}
+
+fn char_cmp(a_char: char, b_char: char) -> Ordering {
+    let a = caseless::default_case_fold_str(&a_char.to_string());
+    let b = caseless::default_case_fold_str(&b_char.to_string());
+
+    let first_char = a.chars().nth(0).unwrap_or('{');
+
+    let order = if a == b && a.len() == 1 && 'a' <= first_char && first_char <= 'z' {
+        if a_char > b_char {
+            Ordering::Less
+        } else if a_char < b_char {
+            Ordering::Greater
+        } else {
+            Ordering::Equal
+        }
+    } else {
+        a.cmp(&b)
+    };
+
+    order
+}
